feat(application-groups): implement application groups
