+++ backend/common/items.go
@@ -352,6 +352,7 @@ func (i Question) GetKey() uuid.UUID {
 type Application struct {
 	ID                  int
 	UUID                uuid.UUID
+	GroupID             uuid.UUID
 	Default             bool
 	Code                string
 	ClientVersion       string
@@ -363,6 +364,17 @@ type Application struct {
 
 // GetKey returns the key for this item
 func (i Application) GetKey() int {
+	return i.ID
+}
+
+// ApplicationGroup contains data for
+type ApplicationGroup struct {
+	ID    uuid.UUID
+	Roles []string
+}
+
+// GetKey returns the key for this item
+func (i ApplicationGroup) GetKey() uuid.UUID {
 	return i.ID
 }
