+++ backend/sqlc/profile-queries.go
@@ -22,12 +22,15 @@ func (q *Queries) ProfileQueries(profileID uuid.UUID) *ProfileQueries {
 }
 
 // GetProfilesForUserIDs retrieves profiles for the specific users.
-func (q *Queries) GetProfilesForUserIDs(ctx context.Context, userIDs []string) ([]common.Profile, error) {
-	profiles, err := q.getProfiles(ctx, userIDs)
+func (aq *ApplicationQueries) GetProfilesForUserIDs(ctx context.Context, userIDs []string) ([]common.Profile, error) {
+	profiles, err := aq.getProfiles(ctx, getProfilesParams{
+		UserID:             userIDs,
+		ApplicationgroupID: aq.groupID,
+	})
 	if err != nil {
 		return nil, err
 	}
-	return lo.Map(profiles, func(i UsersProfile, _ int) common.Profile {
+	return lo.Map(profiles, func(i getProfilesRow, _ int) common.Profile {
 		return common.Profile(i)
 	}), nil
 }
@@ -35,8 +38,9 @@ func (q *Queries) GetProfilesForUserIDs(ctx context.Context, userIDs []string) (
 // SaveProfile creates or overwrites a profile in the database
 func (q *Queries) SaveProfile(ctx context.Context, profile common.Profile) error {
 	return q.saveProfile(ctx, saveProfileParams{
-		Column1: profile.ID,
-		Column2: profile.UserID,
-		Column3: profile.Name,
+		ID:                 profile.ID,
+		UserID:             profile.UserID,
+		Name:               profile.Name,
+		ApplicationgroupID: profile.ApplicationGroupID,
 	})
 }
