+++ queries/applications.sql
@@ -1,10 +1,11 @@
 -- name: getApplications :many
-WITH roles AS (SELECT r.applications_id,
+WITH roles AS (SELECT r.applicationgroups_id,
                       array_agg(DISTINCT r.usergroups_code) as roles
-               FROM applications_usergroups r
-               GROUP BY r.applications_id)
+               FROM applicationgroups_usergroups r
+               GROUP BY r.applicationgroups_id)
 SELECT a.id::int                          AS id,
        a.uuid                             AS uuid,
+       a.group_id                         AS group_id,
        a.code::varchar                    AS code,
        a.default                          AS "default",
        a.client_version,
@@ -14,17 +15,19 @@ SELECT a.id::int                          AS id,
        a.standalone_related_collection_id AS standalone_related_collection_id,
        COALESCE(r.roles, '{}')::varchar[] AS roles
 FROM applications a
-         LEFT JOIN roles r ON a.id = r.applications_id
+         JOIN applicationgroups g ON g.id = a.group_id
+         LEFT JOIN roles r ON g.id = r.applicationgroups_id
 WHERE a.id = ANY ($1::int[])
   AND a.status = 'published';
 
 -- name: listApplications :many
-WITH roles AS (SELECT r.applications_id,
+WITH roles AS (SELECT r.applicationgroups_id,
                       array_agg(DISTINCT r.usergroups_code) as roles
-               FROM applications_usergroups r
-               GROUP BY r.applications_id)
+               FROM applicationgroups_usergroups r
+               GROUP BY r.applicationgroups_id)
 SELECT a.id::int                          AS id,
        a.uuid                             AS uuid,
+       a.group_id                         AS group_id,
        a.code::varchar                    AS code,
        a.default                          AS "default",
        a.client_version,
@@ -34,8 +37,20 @@ SELECT a.id::int                          AS id,
        a.standalone_related_collection_id AS standalone_related_collection_id,
        COALESCE(r.roles, '{}')::varchar[] AS roles
 FROM applications a
-         LEFT JOIN roles r ON a.id = r.applications_id
+         JOIN applicationgroups g ON g.id = a.group_id
+         LEFT JOIN roles r ON g.id = r.applicationgroups_id
 WHERE a.status = 'published';
+
+-- name: getApplicationGroups :many
+WITH roles AS (SELECT r.applicationgroups_id,
+                      array_agg(DISTINCT r.usergroups_code) as roles
+               FROM applicationgroups_usergroups r
+               GROUP BY r.applicationgroups_id)
+SELECT g.id,
+       COALESCE(r.roles, '{}')::varchar[] AS roles
+FROM applicationgroups g
+         LEFT JOIN roles r ON g.id = r.applicationgroups_id
+WHERE g.id = ANY (@id::uuid[]);
 
 -- name: getApplicationIDsForCodes :many
 SELECT p.id, p.code
