+++ queries/user-profiles.sql
@@ -1,9 +1,13 @@
 -- name: getProfiles :many
-SELECT *
-FROM users.profiles
-WHERE user_id = ANY ($1::varchar[]);
+SELECT p.id,
+       p.user_id,
+       p.name,
+       p.applicationgroup_id AS application_group_id
+FROM users.profiles p
+WHERE applicationgroup_id = @applicationgroup_id::uuid
+  AND user_id = ANY (@user_id::varchar[]);
 
 -- name: saveProfile :exec
-INSERT INTO users.profiles (id, user_id, name)
-VALUES ($1::uuid, $2::varchar, $3::varchar)
-ON CONFLICT (id) DO UPDATE SET name = EXCLUDED.name ;
+INSERT INTO users.profiles (id, user_id, name, applicationgroup_id)
+VALUES (@id::uuid, @user_id::varchar, @name::varchar, @applicationgroup_id::uuid)
+ON CONFLICT (id, applicationgroup_id) DO UPDATE SET name = EXCLUDED.name;
